+++ src/main/java/com/tplus/gwland/cat/domain/CategoryDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 @Lazy
 public class CategoryDto {
-	private int catNum;
+	private long catNum;
 	private String areacode;
 	private String cat1;
 	private String cat2;
