+++ src/main/java/com/tplus/gwland/cmm/controller/AbstractController.java
@@ -6,11 +6,11 @@ import java.util.Optional;
 import org.springframework.http.ResponseEntity;
 
 public abstract class AbstractController<T> {
-	public abstract ResponseEntity<Integer> save(T t);
-	public abstract ResponseEntity<Integer> delete(T t);
-	public abstract ResponseEntity<Integer> count();
-	public abstract ResponseEntity<T> getOne(int id);
-	public abstract ResponseEntity<Optional<T>> findById(int id);
-	public abstract ResponseEntity<Boolean> existsById(int id);
+	public abstract ResponseEntity<Long> save(T t);
+	public abstract ResponseEntity<Long> delete(T t);
+	public abstract ResponseEntity<Long> count();
+	public abstract ResponseEntity<T> getOne(long id);
+	public abstract ResponseEntity<Optional<T>> findById(long id);
+	public abstract ResponseEntity<Boolean> existsById(long id);
 	public abstract ResponseEntity<List<T>> findAll();
 }
