+++ src/main/java/com/tplus/gwland/cmm/service/AbstractService.java
@@ -4,12 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 public abstract class AbstractService<T> {
-	public abstract int save(T t);
-	public abstract int delete(T t);
-	public abstract int count();
-	public abstract T getOne(int id);
-	public abstract Optional<T> findById(int id);
-	public abstract boolean existsById(int id);
+	public abstract long save(T t);
+	public abstract long delete(T t);
+	public abstract long count();
+	public abstract T getOne(long id);
+	public abstract Optional<T> findById(long id);
+	public abstract boolean existsById(long id);
 	public abstract List<T> findAll();
 }
