+++ src/main/java/com/tplus/gwland/pce/service/PlaceService.java
@@ -6,5 +6,5 @@ import com.tplus.gwland.pce.domain.Place;
 
 public interface PlaceService{
 	public List<Place> findByContentidAndTitle(String contentid, String title);
-	public int update (String tel, int pceNum);
+	public int update (String tel, long pceNum);
 }
